Add \Tymon\JWTAuth\Providers\AbstractServiceProvider
